+++ cooked_input/tests/test_convertors.py
@@ -6,6 +6,7 @@ pytest tests for cooked_input -- test convertor functions
 Len Wanger, 2017
 """
 
+import decimal
 import sys
 
 if sys.version_info[0] > 2:  # For Python 3
@@ -13,8 +14,8 @@ if sys.version_info[0] > 2:  # For Python 3
 else:
     from StringIO import StringIO
 
-from cooked_input import get_input, get_boolean, get_list, get_date, get_yes_no
-from cooked_input import Convertor, IntConvertor, BooleanConvertor, ListConvertor, DateConvertor, YesNoConvertor
+from cooked_input import get_input, get_boolean, get_list, get_date, get_yes_no, get_money
+from cooked_input import Convertor, IntConvertor, BooleanConvertor, ListConvertor, DateConvertor, YesNoConvertor, DecimalConvertor
 from cooked_input import StripCleaner
 from .utils import redirect_stdin
 
@@ -137,3 +138,43 @@ class TestConvertors(object):
             print(result)
             assert (result == 'no')
         print(ync)
+
+    def test_decimal(self):
+        input_str = u"""
+            10.1
+            10.10
+            10.100
+            """
+
+        dc = DecimalConvertor(precision=2)
+        good_result = decimal.Decimal('10.10')
+        with redirect_stdin(StringIO(input_str)):
+            result = get_input(cleaners=StripCleaner(), convertor=dc)
+            print(result)
+            assert (result == good_result)
+        print(dc)
+
+    def test_get_money(self):
+        input_str = u"""
+            $10.17
+            10.17
+            """
+
+        good_result = decimal.Decimal('10.17')
+
+        with redirect_stdin(StringIO(input_str)):
+            result = get_money(symbol="$")
+            print(result)
+            assert (result == good_result)
+
+    def test_get_money2(self):
+        input_str = u"""
+            $1,000,012.17
+            """
+
+        good_result = decimal.Decimal('1000012.17')
+
+        with redirect_stdin(StringIO(input_str)):
+            result = get_money(symbol="$", separator=",")
+            print(result)
+            assert (result == good_result)